% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-table.R
\name{add_caption}
\alias{add_caption}
\title{Add Caption}
\usage{
add_caption(tblr_obj, title = NA_character_)
}
\arguments{
\item{tblr_obj}{A valid tabler object}

\item{title}{A character scalar with the new title to use}
}
\value{
A tabler object with the new caption as a title
}
\description{
Adds caption to an existing tabler object or sum_tabler object..
}
