% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{add_suppressed_row}
\alias{add_suppressed_row}
\title{Add Suppressed Row}
\usage{
add_suppressed_row(row_dt, tbl_dt)
}
\arguments{
\item{row_dt}{A tibble containing one row of data to be added}

\item{tbl_dt}{A tibble containing table information}
}
\description{
Adds a new row of suppressed data to an existing table data.frame.
}
