% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{add_tabler_deps}
\alias{add_tabler_deps}
\title{Add all dependencies to a tag. Don't forget to set append to TRUE to preserve any existing
dependency}
\usage{
add_tabler_deps(tag)
}
\arguments{
\item{tag}{An HTML tag}
}
\description{
Add all dependencies to a tag. Don't forget to set append to TRUE to preserve any existing
dependency
}
