% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{alias_column_names}
\alias{alias_column_names}
\title{Alias Column Names}
\usage{
alias_column_names(x, alias_list)
}
\arguments{
\item{x}{Character vector}

\item{alias_list}{The named list of aliase contained by a tabler object}
}
\value{
A character vector of the same length of x, where elements have been
replaced by their alias if an alias exists
}
\description{
Takes a character vector of column names (such as dependent variables or
statistical methods) and determines whether an alias has been supplied.  If
so, this function replaces that element
}
