% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{build_absorb_dt}
\alias{build_absorb_dt}
\title{Build Absorb Data}
\usage{
build_absorb_dt(absorb_list)
}
\description{
Create a data.frame that contains information on the absorbed values used in
each estimation.
}
