% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-column.R
\name{check_for_logicals}
\alias{check_for_logicals}
\title{Check for Logicals}
\usage{
check_for_logicals(col_obj)
}
\arguments{
\item{col_obj}{A tabler_column object}
}
\value{
The same tabler_column with new xlevels added to pick up any logical
variables thare found.
}
\description{
Explores the tabler_column object information to determine if there are any
coefficients that are logical values that have not been appropriately
handled. Any that are found are added to the list of factor variables.
}
