% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{clean_errant_codes}
\alias{clean_errant_codes}
\title{Check for Errant Codes}
\usage{
clean_errant_codes(in_kable)
}
\arguments{
\item{in_kable}{Kable object to process}
}
\value{
Kable object
}
\description{
Check for errant text codes where multicolumn contains another multicolumn
}
