% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{coef_to_dt}
\alias{coef_to_dt}
\title{Convert Coef Tibble to Output}
\usage{
coef_to_dt(coef_dt, sig_levels, digits)
}
\arguments{
\item{coef_dt}{A data.frame of coefficient results from a tabler_object}

\item{sig_levels}{A named character vector that gives the thresholds to use
for indicating statistical significance. The names of the vector should provide
the characters or symbols to use in denoting that significance level. This
can be found in tabler_object$theme$sig_level}
}
\description{
Takes a coefficient data.frame from a tabler_object and modifies
the format to create a new data.frame that is formatted in the style used to
display coefficient results in most economics journals.
}
