% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-table.R
\name{combine_xlevels}
\alias{combine_xlevels}
\title{Combine xlevels}
\usage{
combine_xlevels(cols)
}
\arguments{
\item{...}{One or more tabler_column objects}
}
\value{
A named list of character vectors where each name is a factor variable
included in the supplied tabler_objects and the character vector gives all of
the levels associated with that factor.
}
\description{
This function takes a series of tabler_column objects and combines the xlevels
into a single list.  This provides a list of factor variables and all the
levels observed in the columns.
}
