% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_data.R
\name{convert_to_data}
\alias{convert_to_data}
\title{Convert a model output to a dataframe}
\usage{
convert_to_data(model, teststat = "std.error", digits = 3,
  digits_coef = digits, digits_teststat = digits, cutoffs = c(0.1, 0.05,
  0.01), stars = c("*", "**", "***"), N = T, fit = NULL)
}
\arguments{
\item{model}{A single model result}

\item{teststat}{Which test statistic to show ('std.error' standard error,
'p.value' p value, 'statistic' model-specific statistic, \code{NA} for
none)}

\item{digits}{The number of digits to display}

\item{digits_coef}{Override for digits for coefficient}

\item{digits_teststat}{Override for digits for test statistic}

\item{cutoffs}{Levels of significance to display stars for, in descending
order}

\item{stars}{Choices for displayed stars}
}
\description{
This takes each model used in the final table and converts it to a tidy
dataset, formatted and ready to be merged with others in a nice table.
}

