% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{css_class}
\alias{css_class}
\title{Create CSS classes string}
\usage{
css_class(...)
}
\arguments{
\item{...}{Class names to combine}
}
\value{
Combined class string
}
\description{
Create CSS classes string
}
\keyword{internal}
