% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw function}
\usage{
draw(data)
}
\arguments{
\item{data}{A data frame formatted exactly how you want your table to look.  Row names will not be preserved.}
}
\description{
Draws a very simple table inside your plotting window based on data in a data frame.  The functions were
adapted from \code{demo(\link{plotmath})}
}
\details{
The \code{tabler} \code{draw} fucntion will simply plot your table in a plotting window.
This is mainly helpful in conjunction with knitr/pandoc.  It will help you convert tables
to word documents without anything exploding or leaking off the page.  You can control
the table size by adjusting \code{fig.height} in the R-markdown code chunk.
}
\examples{
my_table <- data.frame("A" = c(1, 2), "B" = c(3, 4))

draw(my_table)
}

