% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in module.R
\name{get_default_reactive_domain}
\alias{get_default_reactive_domain}
\title{The Currently Running tabler_app Session}
\usage{
get_default_reactive_domain()
}

\value{
A \code{tabler_app} session object, or \code{NULL}.
}

\description{
Returns the \code{session} object of the \code{tabler_app}
  server function currently executing, or \code{NULL} if none is running.
  Used as the default \code{session} argument of \code{\link{module_server}}.
}

