% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in core.R
\name{get_layout_attributes}
\alias{get_layout_attributes}
\title{Get layout-specific body attributes}
\usage{
get_layout_attributes(layout)
}

\arguments{
\item{layout}{Layout type}
}

\value{
A list of body attributes
}

\description{
Determine body classes and attributes based on layout
}

\keyword{internal}
