% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_layout_structure}
\alias{get_layout_structure}
\title{Get layout-specific page structure}
\usage{
get_layout_structure(layout, navbar, body, footer)
}
\arguments{
\item{layout}{Layout type}

\item{navbar}{Navbar component}

\item{body}{Body content}

\item{footer}{Footer component}
}
\value{
A Shiny tag representing the page structure
}
\description{
Build the HTML structure based on layout type
}
\keyword{internal}
