% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in module.R
\name{getDefaultReactiveDomain}
\alias{getDefaultReactiveDomain}
\title{The Currently Running tablerApp Session}
\usage{
getDefaultReactiveDomain()
}

\value{
A \code{tablerApp} session object, or \code{NULL}.
}

\description{
Returns the \code{session} object of the \code{tablerApp}
  server function currently executing, or \code{NULL} if none is running.
  Used as the default \code{session} argument of \code{\link{moduleServer}}.
}

