% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{horizontalMenu}
\alias{horizontalMenu}
\title{Create a Horizontal Navigation Menu}
\usage{
horizontalMenu(...)
}
\arguments{
\item{...}{Navigation items (menuItem)}
}
\description{
Container for navigation items in horizontal layout
}
