% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{html-tags}
\alias{div}
\alias{span}
\alias{p}
\alias{a}
\alias{img}
\alias{ul}
\alias{li}
\alias{aside}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{hr}
\alias{br}
\title{HTML Tag Builder Functions}
\usage{
div(...)
span(...)
p(...)
a(...)
img(...)
ul(...)
li(...)
aside(...)
h1(...)
h2(...)
h3(...)
h4(...)
h5(...)
h6(...)
hr(...)
br(...)
}

\arguments{
\item{...}{Child elements (strings, other tags) and/or named HTML
  attributes (e.g. \code{class = "foo"}, \code{href = "#"}).}
}

\value{
A \code{tabler.tag} object.
}

\description{
Convenience functions for building common HTML elements. Each function
accepts child content and named attributes as \code{...} arguments and
returns a \code{tabler.tag} object that can be nested, rendered with
\code{as.character()}, or passed to any tabler layout function.
}

\seealso{
\code{\link{tags}} for the full HTML5 tag namespace.
}

