% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler-package.R
\docType{data}
\name{icons}
\alias{icons}
\title{Tabler Icons Data}
\format{
A data frame with 6,019 rows and 3 variables:
\describe{
 \item{root}{The root category of the icon (e.g., "arrow/arrows", "brand", "device", etc.)}
 \item{icon}{The full icon name (e.g., "arrow-up", "arrow-down", "circle-arrow-up", "circle-arrow-up-filled", etc.)}
 \item{filled}{Logical indicating whether the icon is a filled or not.}
}
}
\source{
Derived from the icons from \url{https://github.com/tabler/tabler-icons}
}
\usage{
icons
}
\description{
A dataset containing Tabler icon names and their categories.
}
\examples{
icons[grepl("paw", icons$icon), ]
}
\keyword{datasets}
