% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-boxed.R
\name{layout_boxed}
\alias{layout_boxed}
\title{Build boxed layout structure}
\usage{
layout_boxed(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\arguments{
\item{navbar}{Navbar component (or list with top/side for combo)}

\item{sidebar}{Sidebar component (not typically used when passing list)}

\item{body}{Body content}

\item{footer}{Footer component}
}

\description{
Build the HTML structure for the boxed layout. This layout
has a top navbar.
}

\keyword{internal}
