% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-fluid.R
\name{layout_fluid}
\alias{layout_fluid}
\title{Build fluid layout structure}
\usage{
layout_fluid(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\arguments{
\item{navbar}{Navbar component}

\item{sidebar}{Not used}

\item{body}{Body content}

\item{footer}{Footer content}
}

\description{
Build structure for the fluid layout which uses a top navbar
}

\keyword{internal}
