% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-horizontal.R
\name{layout_horizontal}
\alias{layout_horizontal}
\title{Build horizontal layout structure}
\usage{
layout_horizontal(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\arguments{
\item{navbar}{Navbar component}

\item{sidebar}{Not used}

\item{body}{Body content}

\item{footer}{Footer}
}

\description{
Horizontal layout uses a top header/navigation
}

\keyword{internal}
