% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-navbar.R
\name{layout_navbar}
\alias{layout_navbar}
\title{Build navbar layout structure}
\usage{
layout_navbar(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Top navbar variant with light theme wrapper
}

\keyword{internal}
