% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-navbar-overlap.R
\name{layout_navbar_overlap}
\alias{layout_navbar_overlap}
\title{Build navbar overlap layout structure}
\usage{
layout_navbar_overlap(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Overlap navbar variant - similar to condensed but with text-white page header
}

\keyword{internal}
