% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-navbar-sticky.R
\name{layout_navbar_sticky}
\alias{layout_navbar_sticky}
\title{Build navbar sticky layout structure}
\usage{
layout_navbar_sticky(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Sticky navbar wrapper - transforms vertical navbar to horizontal sticky structure
}

\keyword{internal}
