% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-rtl.R
\name{layout_rtl}
\alias{layout_rtl}
\title{Build rtl layout structure}
\usage{
layout_rtl(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Right-to-left layout variant - transforms vertical navbar to horizontal with brand and theme inside ul
}

\keyword{internal}
