% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-vertical.R
\name{layout_vertical}
\alias{layout_vertical}
\title{Build vertical layout structure}
\usage{
layout_vertical(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Vertical layout with sidebar on the left
}

\keyword{internal}
