% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in layout-vertical-right.R
\name{layout_vertical_right}
\alias{layout_vertical_right}
\title{Build vertical-right layout structure}
\usage{
layout_vertical_right(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}

\description{
Vertical layout with sidebar on the right
}

\keyword{internal}
