% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{list_first}
\alias{list_first}
\title{List First}
\usage{
list_first(dt, ...)
}
\arguments{
\item{dt}{A data.frame}

\item{...}{A series of character variables giving the names of the variables
to be listed first.}
}
\description{
Reorders a data.frame with the specified variables appearing first in order.
The order of unspecified varaibles is maintained after all of the specified
variables have been listed.
}
