% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-column.R
\name{make_column}
\alias{make_column}
\title{Make Column object from single lm result.}
\usage{
make_column(in_result)
}
\arguments{
\item{in_result}{An object containing the results of a statistical estimation}
}
\value{
Results a tabler_column object
}
\description{
Make Column object from single lm result.
}
\examples{
make_column(in_result)
}
