% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeColumn.R
\name{makeColumn}
\alias{makeColumn}
\title{Make Column object from single lm result.}
\usage{
makeColumn(inResult)
}
\arguments{
\item{inResult}{An object containing the results of a statistical estimation}
}
\value{
Results a colRec object
}
\description{
Make Column object from single lm result.
}
\examples{
makeColumn(inResult)
}

