% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in core.R
\name{menu_dropdown}
\alias{menu_dropdown}
\title{Create a Dropdown Menu Item}
\usage{
menu_dropdown(text, icon = NULL, href = NULL, items = list())
}

\arguments{
\item{text}{Title of the dropdown}

\item{icon}{Icon name (optional)}

\item{href}{Link for the dropdown toggle (optional)}

\item{items}{A list of character vectors or tags representing dropdown links; each element can be a character vector of length 2: c(text, href)}
}

\value{
An HTML tag representing the dropdown nav item
}

\description{
Create a nav-item with a dropdown menu using columns like the example
}

