% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in core.R
\name{navbar_menu}
\alias{navbar_menu}
\title{Build a standard navbar/header}
\usage{
navbar_menu(..., brand = NULL, show_theme_button = FALSE)
}

\arguments{
\item{...}{Additional nodes to include (typically `menu_item()` elements)}

\item{brand}{Optional brand for the navbar; either a string (text title),
 `sidebar_brand(...)` or a named list with elements `text` and `img`
 (URL/path).}

\item{show_theme_button}{Whether to show the theme toggle buttons (default: `FALSE``).}
}

\value{
An HTML tag for the header
}

\description{
Create the top header used in boxed layout. This function
accepts raw `li` items (created by `menu_item`) or a `sidebar_brand()`
attached as `title` attribute on a `ul`.
}

