% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-order.R
\name{order_variables}
\alias{order_variables}
\title{Order Variables}
\usage{
order_variables(coefs, var_names, xlevels)
}
\description{
This function will return an ordering of the variables as they will appear
in the table
}
