% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-table.R
\name{+.tabler_object}
\alias{+.tabler_object}
\title{Add New Results to Existing Table}
\usage{
\method{+}{tabler_object}(tblr_obj, new_object)
}
\arguments{
\item{tblr_obj}{A tabler_object}

\item{new_object}{The result object to be added.}
}
\value{
A new tabler_object
}
\description{
Allows user to add an additional column of results to a pre-existing
tabler_object
}
