% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print_html}
\alias{print_html}
\title{Print as an HTML Table}
\usage{
print_html(in_table)
}
\arguments{
\item{in_table}{tabler_object for printing}
}
\description{
Prints a tablerObject as a properly formatted HTML table.
}
\examples{
print_html(in_table)
}

