% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print_latex}
\alias{print_latex}
\title{Print as a LaTeX Table.}
\usage{
print_latex(myTable)
}
\arguments{
\item{myTable}{tablerObject for printing}
}
\description{
Prints a tablerObject as a properly formatted latex table.
}
\examples{
print_latex(myTable)
}

