% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print_markdown}
\alias{print_markdown}
\title{Print as a Markdown Table.}
\usage{
print_markdown(in_tabler)
}
\arguments{
\item{in_tabler}{tablerObject for printing}
}
\description{
Prints a tablerObject as a properly formatted R Markdown table.
}
\examples{
print_markdown(in_tabler)
}

