% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-functions.R
\name{process_header_alias}
\alias{process_header_alias}
\title{Process Header Alias}
\usage{
process_header_alias(nm, alias_list)
}
\arguments{
\item{nm}{Character scalar with name of column to process}

\item{alias_list}{Named character vector where name is the column header
(potentially) and the value is the alias.}
}
\value{
A character scalar
}
\description{
Process the header of a summary table to see if any of the columns have a
specified alias.
}
\details{
This is an internal function
}
