% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-tabler.R
\name{return_last_section}
\alias{return_last_section}
\title{Return Last Section}
\usage{
return_last_section(x)
}
\arguments{
\item{x}{Character scalar providing the order of the tabler table.
Normally stored as tblr_obj$theme$order.}
}
\value{
A single character giving the group that precdes where the
line will be added. If the line is to be added at the very top of
the table, will return NA.
}
\description{
Takes a subset of the order specification, strips out all of the
line elements, and returns the section that preceded the last
position.  This will be used to identify the row number that
the line will be added below.
}
