% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-table.R
\name{set_osa}
\alias{set_osa}
\title{Set Omit-Suppress-Alias (OSA)}
\usage{
set_osa(osa, omit = NULL, suppress = NULL, alias = NULL)
}
\arguments{
\item{omit}{Character vector of variable names to be omitted}

\item{suppress}{Character vector of variable names to be suppressed. Only factor
variables can be suppressed.}

\item{alias}{Named character vector of variables to rename.  The name portion of the
vector should be the existing variable name and the value should be its new name.}

\item{to}{A tabler_object}
}
\value{
An updated tabler_object
}
\description{
Replaces an existing OSA specification with the terms supplied to this function.
If any one of the elments is not included, the any existing values of that term
will be maintained.  To erase the existing elements of a term, set its value to NA.
}
