% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sidebarMenu}
\alias{sidebarMenu}
\title{Create a Navigation Menu}
\usage{
sidebarMenu(...)
}
\arguments{
\item{...}{Navigation items (tabler_nav_item)}
}
\description{
Container for navigation items in sidebar
}
