% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_format.R
\name{stat_format}
\alias{stat_format}
\title{Combine Calculated Statistics and Symbols for Publication and Dissemination}
\usage{
stat_format(.data, .recipe, .name = NA, .digits = NA)
}
\arguments{
\item{.data}{A data frame of calculated statistics.}

\item{.recipe}{A recipe used to create a new column of formatted statistics
(optionally combined with symbols) from existing columns of statistical
results. The recipe must be in the form of a quoted string. It may contain
any combination of column names, spaces, and characters. For example:
"n (percent)" or "percent (lcl - ucl)".}

\item{.name}{An optional name to assign to the column created by the recipe.
The default name is "formatted_stats"}

\item{.digits}{The number of decimal places to display.}
}
\value{
A tibble
}
\description{
The stat_format function is intended to make it quick and easy to
format calculated statistics returned by R statistical functions and
combine them with symbols so that they may more easily
be inserted into tables intended for publication. For example, a proportion,
a lower confidence limit, and an upper confidence limit spread across three
columns could be formatted to two decimal places and combined with
spaces, parentheses, and a hyphen as "24.00 (21.00 - 27.00)."
}
\examples{
library(dplyr)
library(tabler)

data(study)

results <- study \%>\%
  filter(!is.na(sex)) \%>\%
  count(sex) \%>\%
  mutate(percent = n / sum(n) * 100)

print(results)

# A tibble: 2 × 3
  sex        n percent
  <fct>  <int>   <dbl>
1 Female    57    57.6
2 Male      42    42.4

results \%>\%
  stat_format("n (percent\%)")

# A tibble: 2 × 4
  sex        n percent formatted_stats
  <fct>  <int>   <dbl> <chr>
1 Female    57    57.6 57 (57.5757575757576\%)
2 Male      42    42.4 42 (42.4242424242424\%)

results \%>\%
  stat_format("n (percent\%)", .digits = 2)

# A tibble: 2 × 4
  sex        n percent formatted_stats
  <fct>  <int>   <dbl> <chr>
1 Female    57    57.6 57 (57.58\%)
2 Male      42    42.4 42 (42.42\%)

result \%>\%
  stat_format("n (percent\%)", .name = "New Column Name", .digits = 2)

# A tibble: 2 × 4
  sex        n percent `New Column Name`
  <fct>  <int>   <dbl> <chr>
1 Female    57    57.6 57 (57.58\%)
2 Male      42    42.4 42 (42.42\%)


}
