% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{style_box}
\alias{style_box}
\title{Print a formatted table object as Unicode box}
\usage{
style_box(formatted, caption = NULL, space = c(1, 1))
}
\arguments{
\item{formatted}{A formatted table or list of formatted tables.}

\item{caption}{Caption(s) for the table(s).}

\item{space}{Numeric vector of length 2 for spacing before and after output.}
}
\value{
Invisibly returns the formatted object.
}
\description{
Print a formatted table object as Unicode box
}
\examples{
tbl <- format_table(head(mtcars))
style_box(tbl)

}
