% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{style_latex}
\alias{style_latex}
\title{Print a formatted table object as LaTeX code}
\usage{
style_latex(formatted, caption = NULL, space = c(1, 1))
}
\arguments{
\item{formatted}{A formatted table or list of formatted tables.}

\item{caption}{Optional caption(s) for the table(s).}

\item{space}{Numeric vector of length 2 for spacing before and after output.}
}
\value{
Invisibly returns the formatted object.
}
\description{
Print a formatted table object as LaTeX code
}
\examples{
tbl <- format_table(head(mtcars))
style_latex(tbl)

# For a list of formatted tables
tbl_list <- list(
  format_table(head(mtcars)),
  format_table(head(iris))
)
style_latex(tbl_list, caption = c("Mtcars Table", "Iris Table"))

}
