% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{sum_tabler}
\alias{sum_tabler}
\title{Function to generate Latex table of summary statistics}
\usage{
sum_tabler(in_data, func_list = list("mean", "sd", "min", "max"),
  na.rm = FALSE, omit = NULL, alias = NULL)
}
\arguments{
\item{in_data}{Data.frame to be summarized}

\item{func_list}{List contaning function names as strings (default is list('mean', 'sd', 'min', 'max'))}

\item{na.rm}{Boolean indicating whether observations with missing values should be dropped (default = FALSE)}
}
\value{
sum_tabler object
}
\description{
Function to generate Latex table of summary statistics
}
