% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{summarize_single_var}
\alias{summarize_single_var}
\title{Summarize a Single Variable}
\usage{
summarize_single_var(x, func_list, na.rm)
}
\arguments{
\item{x}{A numeric vector}

\item{func_list}{Character vector with names of functions to evaluate}

\item{na.rm}{Boolean}
}
\description{
Implements the calculations from summarize_variables.
}
\details{
This is an internal function.
}
