% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize Variables}
\usage{
summarize_variables(nm, data, func_list, na.rm, xlevels)
}
\arguments{
\item{nm}{Character scalar of a variable name from the data.frame}

\item{data}{Data.frame containing variable `nm`.}

\item{func_list}{List of functions to use to evaluate variable expressed
as a character vector}

\item{na.rm}{Boolean indicating if NA values should be removed before running
functions}

\item{xlevels}{List with named vectors of factor levels}
}
\value{
A data.frame with the summary table rows for the variable
}
\description{
Takes the name of a variable in the data.frame supplied to sum_tabler
and calculates the summary statistics for that variable.
}
\details{
This is an internal function.
}
