% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Function to generate Latex table of summary statistics}
\usage{
summaryTable(inData, funcList = list("mean", "sd", "min", "max"),
  dropAllNA = FALSE)
}
\description{
Function to generate Latex table of summary statistics
}

