% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-functions.R
\name{sumtabler2kable}
\alias{sumtabler2kable}
\title{Sum Tabler To Kable}
\usage{
sumtabler2kable(stbl_obj, format = NULL)
}
\arguments{
\item{stbl_obj}{A sum_tabler object}

\item{format}{Character vector providing the format of the table to be
generated (default = NULL)}
}
\value{
A kable object, modified by kableExtra
}
\description{
Take a `sum_tabler` object and convert it into a kable.
}
