% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{suppress_compress}
\alias{suppress_compress}
\title{Suppress Compress}
\usage{
suppress_compress(suppress_var, dt)
}
\arguments{
\item{suppress_var}{Character scalar giving the name of a factor variable to be suppressed}

\item{dt}{The tibble prepared to be printed}
}
\description{
Takes rows from the coefficient table that are included in the suppression list
and generates a single replacement row for each variable that reports whether
that factor variable was included in the estimation reported in each column.
}
