% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in tabler-package.R
\name{tabler-package}
\alias{tabler-package}
\title{Create Dashboards with 'Tabler'}
\description{
Provides functions to build interactive dashboards using the 'Tabler', making it
 easy to create professional-looking web applications. 'Tabler' is fully responsive and compatible with all modern
 browsers. Offers customizable layouts and components built with 'HTML5' and 'CSS3'. The underlying 'Tabler'
 (<https://github.com/tabler/tabler>) and 'Tabler Icons' (<https://github.com/tabler/tabler-icons>)
 were pre-built from source to eliminate the need for 'Node.js' and 'NPM' on package installation.
}

