% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-table.R
\name{tabler}
\alias{tabler}
\title{Make a tabler object}
\usage{
tabler(..., title = NA_character_, notes = NA_character_,
  number = NA_character_, latex_label = NA_character_,
  alias = NA_character_, suppress = NA_character_,
  omit = NA_character_, gof_list = c(N = "N", r.squared = "R-squared",
  adj.r.squared = "Adj. R-squared"))
}
\description{
This function produces the underlying tabler object.
}
