% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tabler-objects.R
\name{tabler2kable}
\alias{tabler2kable}
\title{Tabler To Kable}
\usage{
tabler2kable(tblr_obj, format = NULL)
}
\arguments{
\item{tblr_obj}{A tabler object}

\item{format}{Character scalar indicating the format to use for the kable
(default is NULL)}
}
\description{
Convert a tabler object into a knitr::kable object.
}
