% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tabler_body}
\alias{tabler_body}
\title{Create a Dashboard Body}
\usage{
tabler_body(..., class = NULL)
}
\arguments{
\item{...}{Content to include in the body}

\item{class}{Additional CSS classes}
}
\description{
Container for dashboard content
}
