% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tabler_footer}
\alias{tabler_footer}
\title{Create a Dashboard Footer}
\usage{
tabler_footer(left = NULL, right = NULL)
}
\arguments{
\item{left}{Left-aligned content}

\item{right}{Right-aligned content}
}
\description{
Footer for the dashboard
}
