% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tabler_icon}
\alias{tabler_icon}
\title{Create an Icon}
\usage{
tabler_icon(name, library = "tabler", class = NULL)
}
\arguments{
\item{name}{Icon name}

\item{library}{Icon library: "tabler", "bootstrap", "feather"}

\item{class}{Additional CSS classes}
}
\description{
Display an icon from Tabler Icons or other icon libraries
}
