% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabler_theme.R
\name{tabler_theme}
\alias{tabler_theme}
\title{Set tabler theme elements}
\usage{
tabler_theme(...)
}
\arguments{
\item{show_dep_var}{= TRUE}

\item{show_method}{= TRUE}

\item{digits}{= c(3,2)}

\item{sig_levels}{= c('***' = 0.01, '**' = 0.05, '*' = 0.1)}
}
\value{
tablerTheme object
}
\description{
Use this function to modify theme settings
}
\details{
This function allows the user to specify the details of how tables are displayed.  The
tabler theme is designed to operate in the same way as themes in ggplot2.  The theme function
can be customized by the user once and then easily applied to all subsequent tables.  This
allows a consistent presentation among tables that does not require passing the same options
into each tabler function.
}

