% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tabler_value_box}
\alias{tabler_value_box}
\title{Create a Value Box}
\usage{
tabler_value_box(value, title, icon = NULL, color = "primary", width = 3)
}
\arguments{
\item{value}{Main value to display}

\item{title}{Title/label for the value}

\item{icon}{Icon name (optional)}

\item{color}{Color theme}

\item{width}{Column width (1-12)}
}
\description{
Display a key metric or value prominently
}
