% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tablerAlert}
\alias{tablerAlert}
\title{Create an Alert/Notification}
\usage{
tablerAlert(..., type = "info", dismissible = FALSE, title = NULL)
}
\arguments{
\item{...}{Alert content}

\item{type}{Alert type: "info", "success", "warning", "danger"}

\item{dismissible}{Whether alert can be dismissed}

\item{title}{Alert title (optional)}
}
\description{
Display important messages to users
}
