% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerBody}
\alias{tablerBody}
\title{Create a Dashboard Body}
\usage{
tablerBody(..., class = NULL)
}
\arguments{
\item{...}{Content to include in the body}

\item{class}{Additional CSS classes}
}
\description{
Container for dashboard content
}
