% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tablerButton}
\alias{tablerButton}
\title{Create a Button}
\usage{
tablerButton(
  label,
  onclick = NULL,
  color = "primary",
  size = "md",
  outline = FALSE,
  icon = NULL,
  ...
)
}
\arguments{
\item{label}{Button text}

\item{onclick}{JavaScript to execute on click}

\item{color}{Button color theme}

\item{size}{Button size: "sm", "md", "lg"}

\item{outline}{Use outline style}

\item{icon}{Icon to include}

\item{...}{Additional HTML attributes}
}
\description{
Create interactive buttons with Tabler styling
}
