% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerFooter}
\alias{tablerFooter}
\title{Create a Dashboard Footer}
\usage{
tablerFooter(left = NULL, right = NULL)
}
\arguments{
\item{left}{Left-aligned content}

\item{right}{Right-aligned content}
}
\description{
Footer for the dashboard
}
