% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerNavbar}
\alias{tablerNavbar}
\title{Create a Dashboard Navbar}
\usage{
tablerNavbar(title = NULL, brand_image = NULL, ...)
}
\arguments{
\item{title}{Brand title/text}

\item{brand_image}{URL or path to brand image}

\item{...}{Additional navbar items}
}
\description{
Top navigation bar for the dashboard
}
