% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerTabItem}
\alias{tablerTabItem}
\title{Create a Tab Item}
\usage{
tablerTabItem(tabName, ...)
}
\arguments{
\item{tabName}{Unique identifier for the tab (must match menuItem tabName)}

\item{...}{Content for this tab}
}
\description{
Individual tab panel content
}
