% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerTabItems}
\alias{tablerTabItems}
\title{Create Tab Items Container}
\usage{
tablerTabItems(...)
}
\arguments{
\item{...}{Tab items created with tablerTabItem()}
}
\description{
Container for multiple tab panels in tabbed layouts
}
