% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{tag_list}
\alias{tag_list}
\title{Create an Ordered Collection of Tags}
\usage{
tag_list(...)
}

\arguments{
\item{...}{Tags or other HTML content.}
}

\value{
A tabler.tag.list object.
}

\description{
Create an Ordered Collection of Tags
}

