% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{tagAppendAttributes}
\alias{tagAppendAttributes}
\title{Append Attributes to an Existing Tag}
\usage{
tagAppendAttributes(tag, ...)
}

\arguments{
\item{tag}{A tabler.tag object.}

\item{...}{Named attributes to add/overwrite. The \code{class} attribute is
  appended (space-separated) rather than replaced.}
}

\value{
The modified tag.
}

\description{
Append Attributes to an Existing Tag
}

