% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{tagList}
\alias{tagList}
\title{Create an Ordered Collection of Tags}
\usage{
tagList(...)
}

\arguments{
\item{...}{Tags or other HTML content.}
}

\value{
A tabler.tag.list object.
}

\description{
Create an Ordered Collection of Tags
}

