% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{tags}
\alias{tags}
\title{HTML Tag Namespace}
\usage{
tags
}
\description{
A named list of tag-builder functions covering the full HTML5
  element set.  Usage: \code{tags$section(class = "foo", "content")}.
}
