% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{tblr_fun}
\alias{tblr_fun}
\title{This function runs a series of functions, in func_list, on a single variable, in_vec.}
\usage{
tblr_fun(x, func_list = NA, ...)
}
\description{
This function runs a series of functions, in func_list, on a single variable, in_vec.
}
