% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{tblr_run}
\alias{tblr_run}
\title{This function runs a single function, f, on a vector of data, in_vec.
This is used by tblr_fun.}
\usage{
tblr_run(f, in_vec)
}
\description{
This function runs a single function, f, on a vector of data, in_vec.
This is used by tblr_fun.
}
