% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_color}
\alias{validate_color}
\title{Validate CSS color values}
\usage{
validate_color(color)
}
\arguments{
\item{color}{Color value to validate}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate CSS color values
}
\keyword{internal}
