% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_size}
\alias{validate_size}
\title{Validate CSS size values}
\usage{
validate_size(size)
}
\arguments{
\item{size}{Size value to validate}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate CSS size values
}
\keyword{internal}
