% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_tab_name}
\alias{validate_tab_name}
\title{Validate tab name format}
\usage{
validate_tab_name(name)
}
\arguments{
\item{name}{Tab name to validate}
}
\description{
Validate tab name format
}
\keyword{internal}
