% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{as.character.tabler.html}
\alias{as.character.tabler.html}
\title{Convert Raw HTML to Character}
\usage{
\method{as.character}{tabler.html}(x, ...)
}

\arguments{
\item{x}{A \code{"tabler.html"} object created by \code{\link{HTML}}.}

\item{...}{Unused.}
}

\value{
The underlying character string.
}

\description{
Convert Raw HTML to Character
}

