% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{col12}
\alias{col12}
\alias{col11}
\alias{col10}
\alias{col9}
\alias{col8}
\alias{col7}
\alias{col6}
\alias{col5}
\alias{col4}
\alias{col3}
\alias{col2}
\alias{col1}
\title{Shorthand for column of width 12}
\usage{
col12(...)

col11(...)

col10(...)

col9(...)

col8(...)

col7(...)

col6(...)

col5(...)

col4(...)

col3(...)

col2(...)

col1(...)
}
\arguments{
\item{...}{Arguments passed to `column()`}
}
\description{
Adapted from the 'golem' package for 'Shiny'.
}
