% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{hidden}
\alias{hidden}
\title{Initialize a Tag as Hidden}
\usage{
hidden(...)
}
\arguments{
\item{...}{Tag (or tagList or list of tags) to make invisible.}
}
\value{
The tag (or tags) that were given as an argument, in a hidden state.
}
\description{
Create a tag (or tag list) that is invisible when the page
  first loads. It can be made visible later with \code{\link{toggle}} or
  \code{\link{show}}.
}
\examples{
if (interactive()) {
  ui <- page(
    title = "Hidden Example",
    body = body(
      actionButton("btn", "Click me"),
      hidden(
        p(id = "element", "I was born invisible")
      )
    )
  )

  server <- function(input, output, session) {
    observeEvent(input$btn, {
      show(session, "element")
    })
  }

  tablerApp(ui, server)
}
}
\seealso{
\code{\link{show}}, \code{\link{hide}}, \code{\link{toggle}}
}
