% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in html.R
\name{HTML}
\alias{HTML}
\title{Mark a String as Raw HTML (Do Not Escape)}
\usage{
HTML(text)
}

\arguments{
\item{text}{Character string of literal HTML.}
}

\value{
An object of class \code{"tabler.html"}.
}

\description{
Mark a String as Raw HTML (Do Not Escape)
}

