% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{HTML}
\alias{HTML}
\title{Mark a String as Raw HTML (Do Not Escape)}
\usage{
HTML(text)
}
\arguments{
\item{text}{Character string of literal HTML.}
}
\value{
An object of class \code{"html"}.
}
\description{
Mark a String as Raw HTML (Do Not Escape)
}
