% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-fluid-vertical.R
\name{layout_fluid_vertical}
\alias{layout_fluid_vertical}
\title{Build fluid vertical layout structure}
\usage{
layout_fluid_vertical(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}
\arguments{
\item{navbar}{Navbar or sidebar component}

\item{sidebar}{Not used}

\item{body}{Body content}

\item{footer}{Footer content}
}
\description{
Vertical fluid layout uses a sidebar (<aside>) and page wrapper
}
\keyword{internal}
