% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-navbar-dark.R
\name{layout_navbar_dark}
\alias{layout_navbar_dark}
\title{Build navbar dark layout structure}
\usage{
layout_navbar_dark(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}
\description{
Top navbar variant with dark theme wrapper
}
\keyword{internal}
