% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-navbar-sticky-dark.R
\name{layout_navbar_sticky_dark}
\alias{layout_navbar_sticky_dark}
\title{Build navbar sticky layout structure (dark variant)}
\usage{
layout_navbar_sticky_dark(
  navbar,
  sidebar,
  body,
  footer,
  color = NULL,
  show_theme_button = TRUE
)
}
\description{
Wrapper for `layout_navbar_sticky()` forcing dark theme for the topbar
}
\keyword{internal}
