% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-vertical-transparent.R
\name{layout_vertical_transparent}
\alias{layout_vertical_transparent}
\title{Build vertical transparent layout structure}
\usage{
layout_vertical_transparent(
  navbar,
  sidebar,
  body,
  footer,
  theme = "light",
  color = NULL,
  show_theme_button = TRUE
)
}
\description{
Vertical layout with transparent sidebar
}
\keyword{internal}
