% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeColumn.R
\name{make_column}
\alias{make_column}
\title{Make Column object from single lm result.}
\usage{
make_column(in_result)
}
\arguments{
\item{inResult}{An object containing the results of a statistical estimation}
}
\value{
Results a colRec object
}
\description{
Make Column object from single lm result.
}
\examples{
make_column(inResult)
}

