% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{navbar_menu}
\alias{navbar_menu}
\title{Build a standard navbar/header}
\usage{
navbar_menu(..., brand = NULL)
}
\arguments{
\item{...}{Additional nodes to include (typically `menu_item()` elements)}
}
\value{
A shiny tag for the header
}
\description{
Create the top header used in boxed layout. This function
accepts raw `li` items (created by `menu_item`) or a `sidebar_brand()`
attached as `title` attribute on a `ul`.
}
