% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variableOrder.R
\name{order_variables}
\alias{order_variables}
\title{Order Variables}
\usage{
order_variables(var_names, xlevels, coefs)
}
\description{
This function will return an ordering of the variables as they will appear in the table
}

