% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-functions.R
\name{print_latex}
\alias{print_latex}
\title{Print as a LaTeX Table.}
\usage{
print_latex(in_tabler)
}
\arguments{
\item{in_tabler}{tablerObject for printing}
}
\description{
Prints a tablerObject as a properly formatted latex table.
}
\examples{
print_latex(in_tabler)
}
