% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler-package.R
\docType{package}
\name{tabler-package}
\alias{tabler}
\alias{tabler-package}
\title{Tabler: Create Dashboards with Tabler and Shiny}
\description{
The tabler package provides a simplified way to create beautiful dashboards
using the modern Tabler CSS framework with Shiny. This package offers a
clean, consistent API with essential components for building professional
web applications.
}
\section{Main Functions}{


Core dashboard structure:
\itemize{
  \item \code{\link{tablerPage}}: Main dashboard page
  \item \code{\link{tablerBody}}: Dashboard body container
  \item \code{\link{tablerNavbar}}: Navigation bar
  \item \code{\link{tablerFooter}}: Footer component
}

UI Components:
\itemize{
  \item \code{\link{tablerCard}}: Card component
  \item \code{\link{tablerValueBox}}: Value/metric display boxes
  \item \code{\link{tablerIcon}}: Icon component
  \item \code{\link{tablerProgress}}: Progress bars
  \item \code{\link{tablerAlert}}: Alert/notification messages
  \item \code{\link{tablerButton}}: Styled buttons
}
}

\examples{
if (interactive()) {
}

}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/pachadotdev/tabler}
  \item Report bugs at \url{https://github.com/pachadotdev/tabler/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com}

Other contributors:
\itemize{
  \item  Tabler (Tabler Dashboard) [contributor, copyright holder]
}

}
\keyword{internal}
