% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler-package.R
\docType{package}
\name{tabler-package}
\alias{tabler}
\alias{tabler-package}
\title{tabler: Create Dashboards with 'Tabler' and 'Shiny'}
\description{
Provides functions to build interactive dashboards combining the 'Tabler UI Kit' with 'Shiny', making it easy to create professional-looking web applications. 'Tabler' is fully responsive and compatible with all modern browsers. Offers customizable layouts and components built with 'HTML5' and 'CSS3'. The underlying 'Tabler' (<https:github.com/tabler/tabler>) and 'Tabler Icons' (<https:github.com/tabler/tabler-icons>) were pre-built from source to eliminate the need for 'Node.js' and 'NPM' on package installation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/tabler}
  \item Report bugs at \url{https://github.com/pachadotdev/tabler/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com}

Other contributors:
\itemize{
  \item  Tabler (Tabler Dashboard) [contributor, copyright holder]
}

}
\keyword{internal}
