% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler-package.R
\docType{package}
\name{tabler-package}
\alias{tabler}
\alias{tabler-package}
\title{tabler: Create Dashboards with 'Tabler' and 'Shiny'}
\description{
Create beautiful dashboards using the Tabler CSS framework with Shiny. This package provides R functions to build interactive dashboards with the modern Tabler Bootstrap theme, making it easy to create professional-looking web applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/tabler}
  \item Report bugs at \url{https://github.com/pachadotdev/tabler/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com}

Other contributors:
\itemize{
  \item  Tabler (Tabler Dashboard) [contributor, copyright holder]
}

}
\keyword{internal}
