% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tabler.R
\docType{package}
\name{tabler}
\alias{tabler}
\alias{tabler-package}
\title{tabler: A package for drawing really simple tables in your plotting window}
\description{
Contains the function \code{\link{draw}} for generating plotted tables.
}
\author{
Alathea D Letaw, \email{alathea@zoology.ubc.ca}
}

