% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Create a table based a JSON retrieved from a Google Sheet}
\usage{
tabler(
  sheetId = NULL,
  rowsByPage = 10,
  paginate = TRUE,
  localization = list(locale = NULL, langs = NULL),
  searchable = FALSE,
  searchableFields = c(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Create a table based a JSON retrieved from a Google Sheet
}
