% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Create a table based a JSON retrieved from a Google Sheet}
\usage{
tabler(sheetId = NULL, width = NULL, height = NULL, elementId = NULL)
}
\description{
Create a table based a JSON retrieved from a Google Sheet
}
