% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\docType{package}
\name{tabler}
\alias{tabler}
\alias{tabler-package}
\title{tabler: a wrapper to quickly combine and share tidy regression output}
\description{
The tabler package is a tool used to make simple tables of
  regression output from multiple models that at least partially overlap in
  variables used. It reports coeffecients, test statistics for variables,
  and, if desired, model fit stats. It can output data as either csv or xls
  and is meant to allow analysts to easily and quickly arrange a series of
  regression outputs to share with collaborators.
}

