% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Convert a set of model outputs to a nice table}
\usage{
tabler(model_list, teststat = "p.value", digits = 3, digits_coef = digits,
  digits_teststat = digits, cutoffs = c(0.1, 0.05, 0.01), stars = c("*",
  "**", "***"), intercept_last = T, save = F)
}
\arguments{
\item{model_list}{Names of a set of model object}

\item{teststat}{Which test statistic to show ('std.error' standard error,
'p.value' p value, 'statistic' model-specific statistic, \code{NA} for
none)}

\item{digits}{The number of digits to display}

\item{digits_coef}{Override for digits for coefficient}

\item{digits_teststat}{Override for digits for test statistic}

\item{cutoffs}{Levels of significance to display stars for, in descending
order}

\item{stars}{Choices for displayed stars}

\item{intercept_last}{If true, move intercept to end of output}

\item{file}{Optional: output the file as a csv}
}
\value{
A dataset of merged output, sorted nicely
}
\description{
Convert a set of model outputs to a nice table
}

