% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tablerSidebar}
\alias{tablerSidebar}
\title{Create a Dashboard Sidebar}
\usage{
tablerSidebar(..., theme = "dark")
}
\arguments{
\item{...}{Sidebar content (typically tabler_nav_menu)}

\item{theme}{Sidebar theme: "light" or "dark"}
}
\description{
Vertical sidebar for navigation
}
