% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{tagList}
\alias{tagList}
\title{Create an Ordered Collection of Tags}
\usage{
tagList(...)
}
\arguments{
\item{...}{Tags or other HTML content.}
}
\value{
A tabler.tag.list object.
}
\description{
Create an Ordered Collection of Tags
}
