% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in visibility.R
\name{visibilityFuncs}
\alias{show}
\alias{show_element}
\alias{hide}
\alias{hide_element}
\alias{toggle}
\alias{toggle_element}
\title{Show/Hide an element}
\usage{
show(
  session = get_default_reactive_domain(),
  id = NULL,
  anim = FALSE,
  anim_type = "slide",
  time = 0.5,
  selector = NULL
)
hide(
  session = get_default_reactive_domain(),
  id = NULL,
  anim = FALSE,
  anim_type = "slide",
  time = 0.5,
  selector = NULL
)
toggle(
  session = get_default_reactive_domain(),
  id = NULL,
  anim = FALSE,
  anim_type = "slide",
  time = 0.5,
  selector = NULL,
  condition = NULL
)
}

\arguments{
\item{session}{The \code{session} object passed by \code{\link{tabler_app}}
  to the server function.}

\item{id}{The id of the element/HTML tag.}

\item{anim}{If \code{TRUE} then animate the behaviour.}

\item{anim_type}{The type of animation to use, either \code{"slide"} or \code{"fade"}.}

\item{time}{The number of seconds to make the animation last.}

\item{selector}{CSS selector of the elements to show/hide. Ignored if the
\code{id} argument is given. For example, to select all span elements with
class x, use \code{selector = "span.x"}.}

\item{condition}{An optional argument to \code{toggle}, see 'Details' below.}
}

\description{
Show or hide an HTML element in a running \code{\link{tabler_app}}.\cr\cr
\strong{\code{show}} makes an element visible, \strong{\code{hide}} makes
an element invisible, \strong{\code{toggle}} displays the element if it is
hidden and hides it if it is visible.\cr\cr
\strong{\code{show_element}}, \strong{\code{hide_element}}, and
\strong{\code{toggle_element}} are synonyms for \code{show}/\code{hide}/
\code{toggle}.\cr\cr
If \code{condition} is given to \code{toggle}, that condition is used to
decide whether to show or hide the element: the element is shown when the
condition evaluates to \code{TRUE} and hidden otherwise.
}

\details{
This is a dependency-free adaptation of \pkg{shinyjs}'s
\code{show}/\code{hide}/\code{toggle} functions for \pkg{tabler}, which does
not use or depend on \pkg{shiny}. Instead of relying on a Shiny session,
these functions use the plain-list \code{session} object created by
\code{\link{tabler_app}} and passed as the third argument to the
\code{server} function.
}

\seealso{
\code{\link{tabler_app}}
}

\examples{
if (interactive()) {
  ui <- page(
    title = "Show/Hide Example",
    body = body(
      action_button("btn", "Click me"),
      div(id = "panel", "Watch what happens to me")
    )
  )

  server <- function(input, output, session) {
    observe_event(input$btn, {
      # Change the following line for more examples
      toggle(session, "panel")
    })
  }

  tabler_app(ui, server)
}
\dontrun{
# The function call in the above app can be replaced by any of the
# following examples to produce similar apps
toggle(session, id = "panel")
toggle(session, "panel", TRUE)
toggle(session, "panel", TRUE, "fade", 2)
toggle(session, id = "panel", time = 1, anim = TRUE, anim_type = "slide")
show(session, "panel")
show(session, id = "panel", anim = TRUE)
hide(session, "panel")
hide(session, id = "panel", anim = TRUE)
}

## toggle can be given an optional `condition` argument, which
## determines if to show or hide the element
if (interactive()) {
  ui <- page(
    title = "Conditional Toggle",
    body = body(
      checkbox_input("checkbox", "Show the text", TRUE),
      div(id = "element", "Watch what happens to me")
    )
  )

  server <- function(input, output, session) {
    observe({
      toggle(session, id = "element", condition = input$checkbox)
    })
  }

  tabler_app(ui, server)
}
}

